reorganizes and fixes tests
