+++ test/index.js
@@ -1,11 +1,17 @@
 const assert = require('assert').strict;
 const t = require('apostrophe/test-lib/util.js');
 const fs = require('fs/promises');
-const {
-  createReadStream
-} = require('fs');
 const path = require('path');
-const FormData = require('form-data');
+const {
+  getAppConfig,
+  insertAdminUser,
+  insertPiecesAndPages,
+  deletePiecesAndPages,
+  deleteAttachments,
+  cleanData,
+  getExtractedFiles,
+  extractFileNames
+} = require('./util');
 
 describe('@apostrophecms/import-export', function () {
   let apos;
@@ -53,6 +59,7 @@ describe('@apostrophecms/import-export', function () {
     await cleanData([ tempPath, exportsPath, attachmentPath ]);
   });
 
+  // FIX
   it('should generate a zip file for pieces without related documents', async function () {
     const req = apos.task.getReq();
     const articles = await apos.article.find(req).toArray();
@@ -86,6 +93,7 @@ describe('@apostrophecms/import-export', function () {
     assert.deepEqual(actual, expected);
   });
 
+  // FIX
   it('should generate a zip file for pieces with related documents', async function () {
     const req = apos.task.getReq();
     const articles = await apos.article.find(req).toArray();
@@ -291,6 +299,7 @@ describe('@apostrophecms/import-export', function () {
     assert.deepEqual(actual, expected);
   });
 
+  // FIX
   it('should return duplicates pieces when already existing and override them', async function() {
     const req = apos.task.getReq();
     const articles = await apos.article.find(req).toArray();
@@ -397,6 +406,7 @@ describe('@apostrophecms/import-export', function () {
     assert.deepEqual(actual, expected);
   });
 
+  // FIX
   it('should import page and related documents', async function() {
     const req = apos.task.getReq();
     const page1 = await apos.page.find(req, { title: 'page1' }).toObject();
@@ -450,6 +460,7 @@ describe('@apostrophecms/import-export', function () {
     assert.deepEqual(actual, expected);
   });
 
+  // FIX
   it('should return existing duplicated docs during page import and override them', async function() {
     const req = apos.task.getReq();
     const page1 = await apos.page.find(req, { title: 'page1' }).toObject();
@@ -554,6 +565,7 @@ describe('@apostrophecms/import-export', function () {
     assert.deepEqual(actual, expected);
   });
 
+  // FIX
   it('should not override attachment if associated document is not imported', async function() {
     const req = apos.task.getReq();
     const page1 = await apos.page.find(req, { title: 'page1' }).toObject();
@@ -738,93 +750,6 @@ describe('@apostrophecms/import-export', function () {
     }), true, `expected imported docs 'lastPublishedAt' value to be of '${lastPublishedAt}'`);
   });
 
-  describe('#checkDuplicates', function() {
-    let apos;
-
-    after(async function() {
-      await t.destroy(apos);
-    });
-
-    before(async function() {
-      apos = await t.create({
-        root: module,
-        testModule: true,
-        modules: getAppConfig({
-          '@apostrophecms/express': {
-            options: {
-              session: { secret: 'supersecret' }
-            }
-          },
-          '@apostrophecms/i18n': {
-            options: {
-              defaultLocale: 'en',
-              locales: {
-                en: { label: 'English' },
-                fr: {
-                  label: 'French',
-                  prefix: '/fr'
-                }
-              }
-            }
-          }
-        })
-      });
-
-      importExportManager = apos.modules['@apostrophecms/import-export'];
-    });
-
-    this.beforeEach(async function() {
-      await insertAdminUser(apos);
-      await insertPiecesAndPages(apos);
-    });
-
-    this.afterEach(async function() {
-      await deletePiecesAndPages(apos);
-      await deleteAttachments(apos, attachmentPath);
-    });
-
-    it('should check if documents to import have duplicates in the current locale', async function() {
-      const req = apos.task.getReq({ mode: 'draft' });
-      const frReq = apos.task.getReq({
-        locale: 'fr',
-        mode: 'draft'
-      });
-      const [ user ] = await apos.doc.db.find({ slug: 'user-admin' }).toArray();
-      const enArticles = await apos.article.find(req).toArray();
-
-      const enDocs = enArticles.concat(user);
-      const enDuplicates = await importExportManager.checkDuplicates(req, enDocs);
-
-      const frDocs = enDocs.map((doc) => {
-        return {
-          ...doc,
-          _id: doc._id.replace('en', 'fr'),
-          aposLocale: 'fr:draft'
-        };
-      });
-      const frDuplicates = await importExportManager.checkDuplicates(frReq, frDocs);
-
-      const actual = {
-        enDuplicates: [
-          enDocs.every((doc) => enDuplicates.duplicatedIds.has(doc.aposDocId)),
-          enDuplicates.duplicatedDocs.length
-        ],
-        frDuplicates: [
-          frDocs.every((doc) => frDuplicates.duplicatedIds.has(doc.aposDocId)),
-          frDuplicates.duplicatedIds.has(user.aposDocId),
-          frDuplicates.duplicatedDocs.length
-        ]
-      };
-      const expected = {
-        enDuplicates: [ true, 3 ],
-        frDuplicates: [ false, true, 1 ]
-      };
-
-      assert.deepStrictEqual(actual, expected);
-
-    });
-  });
-
   describe('#getFirstDifferentLocale', function() {
     it('should find among the docs the first locale that is different from the req one', async function() {
       const req = {
@@ -911,623 +836,6 @@ describe('@apostrophecms/import-export', function () {
           aposMode: 'draft'
         }
       ]);
-    });
-  });
-
-  describe('#import - overriding locales integration tests', function() {
-    let req;
-    let notify;
-    let input;
-    let rewriteDocsWithCurrentLocale;
-    let insertDocs;
-
-    this.beforeEach(async function() {
-      req = apos.task.getReq({
-        locale: 'en',
-        body: {},
-        files: {
-          file: {
-            path: '/some/path/to/file',
-            type: mimeType
-          }
-        }
-      });
-      notify = apos.notify;
-      input = gzip.input;
-      rewriteDocsWithCurrentLocale = apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale;
-      insertDocs = apos.modules['@apostrophecms/import-export'].insertDocs;
-
-      await deletePiecesAndPages(apos);
-      await deleteAttachments(apos, attachmentPath);
-    });
-
-    this.afterEach(function() {
-      apos.notify = notify;
-      gzip.input = input;
-      apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = rewriteDocsWithCurrentLocale;
-      apos.modules['@apostrophecms/import-export'].insertDocs = insertDocs;
-    });
-
-    // FIX
-    it('should import pieces with related documents from the extracted export path when provided', async function() {
-      // Since we are mocking this and not really uploading a file, we have to
-      // manually call setExportPathId to establish a mapping to a safe
-      // unique identifier to share with the "browser"
-      const expectedPath = '/custom/extracted-export-path';
-      await importExportManager.setExportPathId(expectedPath);
-
-      req = apos.task.getReq({
-        locale: 'en',
-        body: {
-          exportPathId: await importExportManager.getExportPathId(expectedPath),
-          formatLabel: 'gzip',
-          overrideLocale: true
-        }
-      });
-
-      gzip.input = async exportPath => {
-        assert.equal(exportPath, expectedPath);
-
-        return {
-          docs: [
-            {
-              _id: '4:en:draft',
-              aposMode: 'draft',
-              aposLocale: 'en:draft',
-              title: 'topic1',
-              type: 'topic'
-            }
-          ],
-          attachmentsInfo: []
-        };
-      };
-
-      await importExportManager.import(req);
-    });
-
-    describe('when the site has only one locale', function() {
-      // FIX
-      it('should not rewrite the docs locale nor ask about it when the locale is not different', async function() {
-        gzip.input = async () => {
-          return {
-            docs: [
-              {
-                _id: '4:en:draft',
-                aposMode: 'draft',
-                aposLocale: 'en:draft',
-                title: 'topic1',
-                type: 'topic'
-              }
-            ],
-            attachmentsInfo: []
-          };
-        };
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = (req, docs) => {
-          throw new Error('rewriteDocsWithCurrentLocale should not have been called');
-        };
-        apos.modules['@apostrophecms/import-export'].insertDocs = async (req, { docs }) => {
-          assert.deepEqual(docs, [
-            {
-              _id: '4:en:draft',
-              aposMode: 'draft',
-              aposLocale: 'en:draft',
-              title: 'topic1',
-              type: 'topic'
-            }
-          ]);
-
-          return [];
-        };
-        apos.notify = async (req, message, options) => {
-          if (options?.event?.name === 'import-export-import-locale-differs') {
-            throw new Error('notify should not have been called with event "import-locale-differ"');
-          }
-          return {};
-        };
-
-        await importExportManager.import(req);
-      });
-
-      // FIX
-      it('should rewrite the docs locale without asking about it when the locale is different', async function() {
-        gzip.input = async () => {
-          return {
-            docs: [
-              {
-                _id: '4:fr:draft',
-                aposMode: 'draft',
-                aposLocale: 'fr:draft',
-                title: 'topic1',
-                type: 'topic'
-              }
-            ],
-            attachmentsInfo: []
-          };
-        };
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = (req, docs) => {
-          assert.deepEqual(docs, [
-            {
-              _id: '4:fr:draft',
-              aposMode: 'draft',
-              aposLocale: 'fr:draft',
-              title: 'topic1',
-              type: 'topic'
-            }
-          ]);
-
-          return rewriteDocsWithCurrentLocale(req, docs);
-        };
-        apos.modules['@apostrophecms/import-export'].insertDocs = async (req, { docs }) => {
-          assert.deepEqual(docs, [
-            {
-              _id: '4:en:draft',
-              aposMode: 'draft',
-              aposLocale: 'en:draft',
-              title: 'topic1',
-              type: 'topic'
-            }
-          ]);
-
-          return {
-            duplicatedDocs: [],
-            duplicatedIds: [],
-            failedIds: []
-          };
-        };
-        apos.notify = async (req, message, options) => {
-          if (options?.event?.name === 'import-export-import-locale-differs') {
-            throw new Error('notify should not have been called with event "import-locale-differ"');
-          }
-          return {};
-        };
-
-        await importExportManager.import(req);
-      });
-    });
-
-    describe('when the site has multiple locales', function() {
-      let apos;
-      let importExportManager;
-
-      let req;
-      let notify;
-      let input;
-      let rewriteDocsWithCurrentLocale;
-      let insertDocs;
-
-      after(async function() {
-        await t.destroy(apos);
-      });
-
-      before(async function() {
-        apos = await t.create({
-          root: module,
-          testModule: true,
-          modules: getAppConfig({
-            '@apostrophecms/express': {
-              options: {
-                session: { secret: 'supersecret' }
-              }
-            },
-            '@apostrophecms/i18n': {
-              options: {
-                defaultLocale: 'en',
-                locales: {
-                  en: { label: 'English' },
-                  fr: {
-                    label: 'French',
-                    prefix: '/fr'
-                  }
-                }
-              }
-            }
-          })
-        });
-
-        importExportManager = apos.modules['@apostrophecms/import-export'];
-        importExportManager.removeExportFileFromUploadFs = () => {};
-        importExportManager.remove = () => {};
-
-        await insertAdminUser(apos);
-        await insertPiecesAndPages(apos);
-      });
-
-      this.beforeEach(async function() {
-        req = apos.task.getReq({
-          locale: 'en',
-          body: {},
-          files: {
-            file: {
-              path: '/some/path/to/file',
-              type: mimeType
-            }
-          }
-        });
-        notify = apos.notify;
-        input = gzip.input;
-        rewriteDocsWithCurrentLocale = apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale;
-        insertDocs = apos.modules['@apostrophecms/import-export'].insertDocs;
-
-        await deletePiecesAndPages(apos);
-        await deleteAttachments(apos, attachmentPath);
-      });
-
-      this.afterEach(function() {
-        apos.notify = notify;
-        gzip.input = input;
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = rewriteDocsWithCurrentLocale;
-        apos.modules['@apostrophecms/import-export'].insertDocs = insertDocs;
-      });
-
-      it('should not rewrite the docs locale nor ask about it when the locale is not different', async function() {
-        const req = apos.task.getReq({
-          locale: 'fr',
-          body: {},
-          files: {
-            file: {
-              path: '/some/path/to/file',
-              type: mimeType
-            }
-          }
-        });
-
-        gzip.input = async req => {
-          return {
-            docs: [
-              {
-                _id: '4:fr:draft',
-                aposMode: 'draft',
-                aposLocale: 'fr:draft',
-                title: 'topic1',
-                type: 'topic'
-              }
-            ],
-            attachmentsInfo: []
-          };
-        };
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = () => {
-          throw new Error('rewriteDocsWithCurrentLocale should not have been called');
-        };
-        apos.modules['@apostrophecms/import-export'].insertDocs = async (req, { docs }) => {
-          assert.deepEqual(docs, [
-            {
-              _id: '4:fr:draft',
-              aposMode: 'draft',
-              aposLocale: 'fr:draft',
-              title: 'topic1',
-              type: 'topic'
-            }
-          ]);
-
-          return {
-            duplicatedDocs: [],
-            duplicatedIds: [],
-            failedIds: []
-          };
-        };
-        apos.notify = async (req, message, options) => {
-          if (options?.event?.name === 'import-export-import-locale-differs') {
-            throw new Error('notify should not have been called with event "import-locale-differ"');
-          }
-          return {};
-        };
-
-        await importExportManager.import(req);
-      });
-
-      it('should not rewrite the docs locales nor insert them but ask about it when the locale is different', async function() {
-        gzip.input = async req => {
-          return {
-            docs: [
-              {
-                _id: '4:fr:draft',
-                aposMode: 'draft',
-                aposLocale: 'fr:draft',
-                title: 'topic1',
-                type: 'topic'
-              }
-            ],
-            attachmentsInfo: []
-          };
-        };
-
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = () => {
-          throw new Error('rewriteDocsWithCurrentLocale should not have been called');
-        };
-        apos.modules['@apostrophecms/import-export'].insertDocs = async (req, docs) => {
-          throw new Error('insertDocs should not have been called');
-        };
-        apos.notify = async (req, message, options) => {
-          assert.equal(options.event.name, 'import-export-import-locale-differs');
-        };
-
-        await importExportManager.import(req);
-      });
-
-      it('should rewrite the docs locale when the locale is different and the `overrideLocale` param is provided', async function() {
-        // Since we are mocking this and not really uploading a file, we have to
-        // manually call setExportPathId to establish a mapping to a safe
-        // unique identifier to share with the "browser"
-        const expectedPath = '/custom/extracted-export-path';
-        await importExportManager.setExportPathId(expectedPath);
-
-        const req = apos.task.getReq({
-          locale: 'en',
-          body: {
-            exportPathId: await importExportManager.getExportPathId(expectedPath),
-            formatLabel: 'gzip',
-            overrideLocale: true
-          }
-        });
-
-        gzip.input = async req => {
-          return {
-            docs: [
-              {
-                _id: '4:fr:draft',
-                aposMode: 'draft',
-                aposLocale: 'fr:draft',
-                title: 'topic1',
-                type: 'topic'
-              }
-            ],
-            attachmentsInfo: []
-          };
-        };
-
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = (req, docs) => {
-          assert.deepEqual(docs, [
-            {
-              _id: '4:fr:draft',
-              aposMode: 'draft',
-              aposLocale: 'fr:draft',
-              title: 'topic1',
-              type: 'topic'
-            }
-          ]);
-
-          return rewriteDocsWithCurrentLocale(req, docs);
-        };
-        apos.modules['@apostrophecms/import-export'].insertDocs = async (req, { docs }) => {
-          assert.deepEqual(docs, [
-            {
-              _id: '4:en:draft',
-              aposMode: 'draft',
-              aposLocale: 'en:draft',
-              title: 'topic1',
-              type: 'topic'
-            }
-          ]);
-
-          return [];
-        };
-        apos.notify = async (req, message, options) => {
-          if (options?.event?.name === 'import-export-import-locale-differs') {
-            throw new Error('notify should not have been called with event "import-locale-differ"');
-          }
-          return {};
-        };
-
-        await importExportManager.import(req);
-      });
-    });
-  });
-
-  describe('#overrideDuplicates - overriding locales integration tests', function() {
-    let req;
-    let input;
-    let rewriteDocsWithCurrentLocale;
-    let jobManager;
-
-    this.beforeEach(async function() {
-      req = apos.task.getReq({
-        locale: 'en',
-        body: {
-          formatLabel: 'gzip'
-        }
-      });
-      jobManager = apos.modules['@apostrophecms/job'];
-      input = gzip.input;
-      rewriteDocsWithCurrentLocale = apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale;
-
-      jobManager.success = () => {};
-      jobManager.failure = () => {};
-
-      await deletePiecesAndPages(apos);
-      await deleteAttachments(apos, attachmentPath);
-    });
-
-    this.afterEach(function() {
-      gzip.input = input;
-      apos.modules['@apostrophecms/job'].jobManager = jobManager;
-      apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = rewriteDocsWithCurrentLocale;
-    });
-
-    describe('when the site has only one locale', function() {
-      // FIX
-      it('should not rewrite the docs locale when the locale is not different', async function() {
-        gzip.input = async exportPath => {
-          return {
-            docs: [
-              {
-                _id: '4:en:draft',
-                aposMode: 'draft',
-                aposLocale: 'en:draft',
-                title: 'topic1',
-                type: 'topic'
-              }
-            ],
-            attachmentsInfo: []
-          };
-        };
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = (req, docs) => {
-          throw new Error('rewriteDocsWithCurrentLocale should not have been called');
-        };
-
-        await importExportManager.overrideDuplicates(req);
-      });
-
-      it('should rewrite the docs locale when the locale is different', async function() {
-        gzip.input = async exportPath => {
-          return {
-            docs: [
-              {
-                _id: '4:fr:draft',
-                aposMode: 'draft',
-                aposLocale: 'fr:draft',
-                title: 'topic1',
-                type: 'topic'
-              }
-            ],
-            attachmentsInfo: []
-          };
-        };
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = (req, docs) => {
-          assert.deepEqual(docs, [
-            {
-              _id: '4:fr:draft',
-              aposMode: 'draft',
-              aposLocale: 'fr:draft',
-              title: 'topic1',
-              type: 'topic'
-            }
-          ]);
-
-          return rewriteDocsWithCurrentLocale(req, docs);
-        };
-
-        await importExportManager.overrideDuplicates(req);
-      });
-    });
-
-    describe('when the site has multiple locales', function() {
-      let apos;
-      let importExportManager;
-
-      let input;
-      let rewriteDocsWithCurrentLocale;
-
-      after(async function() {
-        await t.destroy(apos);
-      });
-
-      before(async function() {
-        apos = await t.create({
-          root: module,
-          testModule: true,
-          modules: getAppConfig({
-            '@apostrophecms/express': {
-              options: {
-                session: { secret: 'supersecret' }
-              }
-            },
-            '@apostrophecms/i18n': {
-              options: {
-                defaultLocale: 'en',
-                locales: {
-                  en: { label: 'English' },
-                  fr: {
-                    label: 'French',
-                    prefix: '/fr'
-                  }
-                }
-              }
-            }
-          })
-        });
-
-        importExportManager = apos.modules['@apostrophecms/import-export'];
-        importExportManager.removeExportFileFromUploadFs = () => {};
-        importExportManager.remove = () => {};
-
-        await insertAdminUser(apos);
-        await insertPiecesAndPages(apos);
-      });
-
-      this.beforeEach(async function() {
-        req = apos.task.getReq({
-          locale: 'en',
-          body: {
-            formatLabel: 'gzip'
-          }
-        });
-        input = gzip.input;
-        rewriteDocsWithCurrentLocale = apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale;
-
-        jobManager = apos.modules['@apostrophecms/job'];
-        jobManager.success = () => {};
-        jobManager.failure = () => {};
-
-        await deletePiecesAndPages(apos);
-        await deleteAttachments(apos, attachmentPath);
-      });
-
-      this.afterEach(function() {
-        gzip.input = input;
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = rewriteDocsWithCurrentLocale;
-      });
-
-      it('should not rewrite the docs locale when the locale is not different', async function() {
-        gzip.input = async exportPath => {
-          return {
-            docs: [
-              {
-                _id: '4:en:draft',
-                aposMode: 'draft',
-                aposLocale: 'en:draft',
-                title: 'topic1',
-                type: 'topic'
-              }
-            ],
-            attachmentsInfo: []
-          };
-        };
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = (req, docs) => {
-          throw new Error('rewriteDocsWithCurrentLocale should not have been called');
-        };
-
-        await importExportManager.overrideDuplicates(req);
-      });
-
-      it('should rewrite the docs locale when the locale is different and the `overrideLocale` param is provided', async function() {
-        const req = apos.task.getReq({
-          locale: 'en',
-          body: {
-            formatLabel: 'gzip',
-            overrideLocale: true
-          }
-        });
-
-        gzip.input = async exportPath => {
-          return {
-            docs: [
-              {
-                _id: '4:fr:draft',
-                aposMode: 'draft',
-                aposLocale: 'fr:draft',
-                title: 'topic1',
-                type: 'topic'
-              }
-            ],
-            attachmentsInfo: []
-          };
-        };
-        apos.modules['@apostrophecms/import-export'].rewriteDocsWithCurrentLocale = (req, docs) => {
-          assert.deepEqual(docs, [
-            {
-              _id: '4:fr:draft',
-              aposMode: 'draft',
-              aposLocale: 'fr:draft',
-              title: 'topic1',
-              type: 'topic'
-            }
-          ]);
-
-          return rewriteDocsWithCurrentLocale(req, docs);
-        };
-
-        await importExportManager.overrideDuplicates(req);
-      });
     });
   });
 
@@ -2177,268 +1485,3 @@ describe('@apostrophecms/import-export', function () {
     });
   });
 });
-
-function extractFileNames (files) {
-  return files.map((fullname) => {
-    const regex = /-([\w\d-]+)\./;
-    const [ , name ] = regex.exec(fullname);
-
-    return name;
-  });
-}
-
-async function getExtractedFiles(extractPath) {
-  const docsData = await fs.readFile(
-    path.join(extractPath, 'aposDocs.json'),
-    { encoding: 'utf8' }
-  );
-
-  const attachmentsData = await fs.readFile(
-    path.join(extractPath, 'aposAttachments.json'),
-    { encoding: 'utf8' }
-  );
-
-  const attachmentFiles = await fs.readdir(path.join(extractPath, 'attachments'));
-
-  return {
-    docs: JSON.parse(docsData),
-    attachments: JSON.parse(attachmentsData),
-    attachmentFiles
-  };
-}
-
-async function cleanData(paths) {
-  try {
-    for (const filePath of paths) {
-      const files = await fs.readdir(filePath);
-      for (const name of files) {
-        await fs.rm(path.join(filePath, name), { recursive: true });
-      }
-    }
-  } catch (err) {
-    assert(!err);
-  }
-}
-
-async function deletePiecesAndPages(apos) {
-  await apos.doc.db.deleteMany({ type: /default-page|article|topic|@apostrophecms\/image/ });
-}
-
-async function deleteAttachments(apos, attachmentPath) {
-  await apos.attachment.db.deleteMany({});
-  await cleanData([ attachmentPath ]);
-}
-
-async function insertPiecesAndPages(apos) {
-  const req = apos.task.getReq();
-
-  const formData = new FormData();
-  formData.append(
-    'file',
-    createReadStream(path.join(apos.rootDir, '/public/test-image.jpg'))
-  );
-
-  const jar = await login(apos.http);
-
-  const attachment = await apos.http.post('/api/v1/@apostrophecms/attachment/upload', {
-    body: formData,
-    jar
-  });
-
-  const image1 = await apos.image.insert(req, {
-    ...apos.image.newInstance(),
-    title: 'image1',
-    attachment
-  });
-
-  const topic3 = await apos.topic.insert(req, {
-    ...apos.topic.newInstance(),
-    title: 'topic3'
-  });
-
-  const topic2 = await apos.topic.insert(req, {
-    ...apos.topic.newInstance(),
-    title: 'topic2'
-  });
-
-  const topic1 = await apos.topic.insert(req, {
-    ...apos.topic.newInstance(),
-    title: 'topic1',
-    _topics: [ topic3 ]
-  });
-
-  await apos.article.insert(req, {
-    ...apos.article.newInstance(),
-    title: 'article1',
-    _topics: [ topic2 ],
-    image: attachment
-  });
-
-  const article2 = await apos.article.insert(req, {
-    ...apos.article.newInstance(),
-    title: 'article2',
-    _topics: [ topic1 ]
-  });
-
-  const pageInstance = await apos.http.post('/api/v1/@apostrophecms/page', {
-    body: {
-      _newInstance: true
-    },
-    jar
-  });
-
-  await apos.page.insert(req, '_home', 'lastChild', {
-    ...pageInstance,
-    title: 'page1',
-    type: 'default-page',
-    _articles: [ article2 ],
-    main: {
-      _id: 'areaId',
-      items: [
-        {
-          _id: 'cllp5ubqk001320613gaz2vmv',
-          metaType: 'widget',
-          type: '@apostrophecms/image',
-          aposPlaceholder: false,
-          imageIds: [ image1.aposDocId ],
-          imageFields: {
-            [image1.aposDocId]: {
-              top: null,
-              left: null,
-              width: null,
-              height: null,
-              x: null,
-              y: null
-            }
-          }
-        }
-      ],
-      metaType: 'area'
-    }
-  });
-}
-
-async function login(http, username = 'admin') {
-  const jar = http.jar();
-
-  await http.post('/api/v1/@apostrophecms/login/login', {
-    body: {
-      username,
-      password: username,
-      session: true
-    },
-    jar
-  });
-
-  const loggedPage = await http.get('/', {
-    jar
-  });
-
-  assert(loggedPage.match(/logged in/));
-
-  return jar;
-}
-
-async function insertAdminUser(apos) {
-  const user = {
-    ...apos.user.newInstance(),
-    title: 'admin',
-    username: 'admin',
-    password: 'admin',
-    role: 'admin'
-  };
-
-  await apos.user.insert(apos.task.getReq(), user);
-}
-
-function getAppConfig(modules = {}) {
-  return {
-    '@apostrophecms/express': {
-      options: {
-        session: { secret: 'supersecret' }
-      }
-    },
-    '@apostrophecms/import-export': {},
-    'home-page': {
-      extend: '@apostrophecms/page-type'
-    },
-    'default-page': {
-      extend: '@apostrophecms/page-type',
-      fields: {
-        add: {
-          main: {
-            label: 'Main',
-            type: 'area',
-            options: {
-              widgets: {
-                '@apostrophecms/rich-text': {},
-                '@apostrophecms/image': {},
-                '@apostrophecms/video': {}
-              },
-              importAsRichText: true
-            }
-          },
-          _articles: {
-            label: 'Articles',
-            type: 'relationship',
-            withType: 'article'
-          }
-        }
-      }
-    },
-
-    article: {
-      extend: '@apostrophecms/piece-type',
-      options: {
-        alias: 'article',
-        autopublish: true
-      },
-      fields: {
-        add: {
-          image: {
-            type: 'attachment',
-            group: 'images'
-          },
-          _topics: {
-            label: 'Topics',
-            type: 'relationship',
-            withType: 'topic'
-          }
-        }
-      }
-    },
-
-    topic: {
-      extend: '@apostrophecms/piece-type',
-      options: {
-        alias: 'topic'
-      },
-      fields: {
-        add: {
-          description: {
-            label: 'Description',
-            type: 'string'
-          },
-          main: {
-            label: 'Main',
-            type: 'area',
-            options: {
-              widgets: {
-                '@apostrophecms/rich-text': {}
-              },
-              importAsRichText: true
-            }
-          },
-          _topics: {
-            label: 'Related Topics',
-            type: 'relationship',
-            withType: 'topic',
-            max: 2
-          }
-        }
-      }
-    },
-
-    ...modules
-  };
-}
